+++ src/main/java/com/hydata/intelligence/platform/model/ApplicationChartDsModel.java
@@ -5,33 +5,31 @@ package com.hydata.intelligence.platform.model;
  * @createTime 2018年11月2日上午9:35:33
  */
 public class ApplicationChartDsModel {
-	private Integer id;
+	private long id;
 	
-	private Integer dd_id;//deviceDatastream,设备数据流id
+	private long dd_id;//deviceDatastream,设备数据流id
 	
-	private Integer chart_id;//图表id
-	
-	
-	
-	
-	public Integer getId() {
+	private long chart_id;//图表id
+		
+	public long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
-	public Integer getDd_id() {
+	public long getDd_id() {
 		return dd_id;
 	}
-	public void setDd_id(Integer dd_id) {
+	public void setDd_id(long dd_id) {
 		this.dd_id = dd_id;
-	}	
-	public Integer getChart_id() {
+	}
+	public long getChart_id() {
 		return chart_id;
 	}
-	public void setChart_id(Integer chart_id) {
+	public void setChart_id(long chart_id) {
 		this.chart_id = chart_id;
 	}
+	
 	
 }
