+++ src/main/java/com/hydata/intelligence/platform/repositories/ApplicationAnalysisRepository.java
@@ -11,8 +11,8 @@ import com.hydata.intelligence.platform.dto.ApplicationAnalysis;
  * @author pyt
  * @createTime 2018年10月24日下午2:09:33
  */
-public interface ApplicationAnalysisRepository extends JpaRepository<ApplicationAnalysis, Integer> {
+public interface ApplicationAnalysisRepository extends JpaRepository<ApplicationAnalysis, Long> {
 	@Query("select aa from ApplicationAnalysis aa where aa.applicationId = ?1")
-	Optional<ApplicationAnalysis> findByApplicationId(Integer application_id);
+	Optional<ApplicationAnalysis> findByApplicationId(long application_id);
 }
