+++ src/main/java/com/hydata/intelligence/platform/repositories/DatastreamModelRepository.java
@@ -21,16 +21,16 @@ import static org.hibernate.jpa.QueryHints.HINT_COMMENT;
  * @createTime 2018年10月24日下午2:05:47
  */
 @Repository
-public interface DatastreamModelRepository extends JpaRepository<DatastreamModel, Integer> ,JpaSpecificationExecutor<DatastreamModel>{
+public interface DatastreamModelRepository extends JpaRepository<DatastreamModel, Long> ,JpaSpecificationExecutor<DatastreamModel>{
 	@Query("select d from DatastreamModel d where d.productId = ?1 and d.name = ?2")
-	List<DatastreamModel> findByProductIdAndName(Integer product_id,String name);
+	List<DatastreamModel> findByProductIdAndName(long product_id,String name);
 	
 	@QueryHints(value = {@QueryHint(name = HINT_COMMENT ,value= "a query for pageable")})
 	@Query("select d from DatastreamModel d where d.productId = ?1")
-	Page<DatastreamModel> queryByProductId(Integer product_id,Pageable pageable);
+	Page<DatastreamModel> queryByProductId(long product_id,Pageable pageable);
 
 	@Query("select d from DatastreamModel d where d.productId = ?1")
-	List<DatastreamModel> findByProductId(Integer product_id);
+	List<DatastreamModel> findByProductId(long product_id);
 
 }
