+++ src/main/java/com/hydata/intelligence/platform/repositories/DeviceTriggerRepository.java
@@ -21,24 +21,24 @@ import com.hydata.intelligence.platform.dto.DeviceTrigger;
  * @author pyt
  * @createTime 2018年10月24日下午2:15:54
  */
-public interface DeviceTriggerRepository extends JpaRepository<DeviceTrigger, Integer> {
+public interface DeviceTriggerRepository extends JpaRepository<DeviceTrigger, Long> {
 	@Modifying
 	@Transactional
 	@Query("delete from DeviceTrigger dt where dt.triggerId = ?1")
-	int deleteByTriggerId(Integer triggerId);
+	int deleteByTriggerId(long triggerId);
 	
 	@Query("select dt from DeviceTrigger dt where dt.triggerId = ?1")
-	List<DeviceTrigger> findByTriggerId(Integer triggerId);
+	List<DeviceTrigger> findByTriggerId(long triggerId);
 	
 	@Query("select dt from DeviceTrigger dt where dt.device_sn = ?1")
 	List<DeviceTrigger> findByDeviceSn(String device_sn);
 	
 	@Query("select dt from DeviceTrigger dt where dt.device_sn = ?1 and dt.triggerId = ?2")
-	Optional<DeviceTrigger> findByDeviceSnAndTriggerId(String device_sn,Integer triggerId);
+	Optional<DeviceTrigger> findByDeviceSnAndTriggerId(String device_sn,long triggerId);
 	
 	@QueryHints(value = {@QueryHint(name = HINT_COMMENT ,value= "a query for pageable")})
 	@Query("select dt from DeviceTrigger dt where dt.triggerId = ?1")
-	Page<DeviceTrigger> queryByTriggerId(Integer trigger_id,Pageable pageable);
+	Page<DeviceTrigger> queryByTriggerId(long trigger_id,Pageable pageable);
 	
 	@QueryHints(value = {@QueryHint(name = HINT_COMMENT ,value= "a query for pageable")})
 	@Query("select dt from DeviceTrigger dt where dt.device_sn = ?1")
