+++ src/main/java/com/hydata/intelligence/platform/repositories/OperationLogsRepository.java
@@ -11,8 +11,8 @@ import com.hydata.intelligence.platform.dto.OperationLogs;
  * @author pyt
  * @createTime 2018年10月24日下午2:17:12
  */
-public interface OperationLogsRepository extends JpaRepository<OperationLogs, Integer> {
+public interface OperationLogsRepository extends JpaRepository<OperationLogs, Long> {
 	@Query("select o from OperationLogs o where o.userId = ?1 and o.msg like concat('%' ,?2,'%')")
-	List<OperationLogs> findByUserIdAndKeyWord(Integer userId,String keyWord);
+	List<OperationLogs> findByUserIdAndKeyWord(long userId,String keyWord);
 }
