+++ src/main/java/com/hydata/intelligence/platform/repositories/TriggerLogsRepository.java
@@ -8,7 +8,7 @@ import com.hydata.intelligence.platform.dto.TriggerLogs;
  * @author pyt
  * @createTime 2018年10月24日下午2:21:15
  */
-public interface TriggerLogsRepository extends JpaRepository<TriggerLogs, Integer> {
+public interface TriggerLogsRepository extends JpaRepository<TriggerLogs, Long> {
 
 }
